refactor: simplify provider actors and Rust idioms
Collapse stateless actor handlers into associated-function calls and streamline small Rust idioms across config, errors, DB, and provider clients.

This also normalizes multiplexing branches and microsecond timing logs without changing behavior.

+++ src/providers/codex/client/client.rs
@@ -88,7 +88,7 @@ impl CodexClient {
                 .await?
                 .ok_or(CodexError::NoAvailableCredential)?;
 
-            let waited_us = start.elapsed().as_micros() as u64;
+            let waited_us = start.elapsed().as_micros();
             info!(
                 waited_us,
                 id = lease.id,
@@ -238,7 +238,7 @@ impl CodexClient {
                 .await?
                 .ok_or(CodexError::NoAvailableCredential)?;
 
-            let waited_us = start.elapsed().as_micros() as u64;
+            let waited_us = start.elapsed().as_micros();
             info!(
                 waited_us,
                 id = lease.id,
